backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.34.0
PyPDF2==3.0.1
python-dotenv==1.0.0
werkzeug==3.0.3